Separated BlogT Monad into its own file and renaming the main to Main.hs
