+++ examples/package.yaml
@@ -31,7 +31,7 @@ when:
 
 executables:
   blog-example:
-    main: Blog.hs
+    main: Main.hs
 
 flags:
   werror:
